Rename express-config to express
express-config.js is already inside the config
folder, removing redundancy.

routes.md's contents has been incorporated into
the index files for each api.
