+++ server/server.js
@@ -1,7 +1,7 @@
 // Require for use with ES6
 import 'babel-polyfill';
 
-import configExpress from './config/express-config';
+import configExpress from './config/express';
 import configRoutes from './routes';
 import express from 'express';
 import db from './database/database';
